Separate staticmap class from utils class file
