+++ src/Extensions/DateTimeExtensions.cs
@@ -25,5 +25,15 @@
             var convertedTime = TimeZoneInfo.ConvertTimeFromUtc(dt, tzInfo);
             return convertedTime;
         }
+
+        /// <summary>
+        /// Get Unix timestamp from current date time
+        /// </summary>
+        /// <param name="now">Date and time to get unix variation from</param>
+        /// <returns>Returns Unix timestamp</returns>
+        public static double GetUnixTimestamp(this DateTime now)
+        {
+            return now.Subtract(new DateTime(1970, 1, 1)).TotalSeconds;
+        }
     }
 }
